chore: rename files from modal context to editview context
